Switch Var Header to binary format
Resolves https://github.com/splunk/stef/issues/25

This eliminates dependency on JSON. All var header is now
in custom binary format, which is fast and compact.

I changed UserData to be map[string]string instead of map[string]any.
This makes the format simpler and is good enough for UserData,
we don't need to support arbitrary types and complicate the
format of the header.

I also updated STEF/gRPC implementation to correctly use the binary
format for schema. It was missed in https://github.com/splunk/stef/pull/23

Note that stefgen command continues to use JSON format, since we
need a human readable format as an input to generator.