Add custom schema serialization format (#23)
Resolves https://github.com/splunk/stef/issues/9

Replaced schema encoding on the wire by a simple binary format instead of JSON.
The format has not much extensibility built-in but any change to the schema
descriptor format is going to be a breaking change anyway, so this simple
approach that is fast and compact works fine for our needs.

For Otel/STEF example schema we get the following byte sizes:

```
      Uncompressed   Zstd
JSON      1877        498
Binary     579        308
Diff:     -69%       -38%
```

Fuzzing discovered a few deserialization bugs, which I fixed and then was
run without failures for another hour.

Alternates Considered: Msgpack, CBOR, Protobuf. All of these would be extra
dependencies that are not necessarily available as a core library in every
language, which would complicate STEF implementation. This simple binary
format is trivial to implement in any language and avoids bringing a
fairly complicated dependency.