Auto-generate randomized Write/Read tests
This helps cover a wider variety of encoding/decoding
paths. For example, it correctly hits an existing bug
with dictionary resets. The expected failing cases
are disabled until https://github.com/splunk/stef/pull/44
is merged.

The test is randomized with non-deterministic starting
seed, but runs deterministically after the seed and prints
the seed value on failures, allowing to reproduce the
failing case locally.