Refactor Reader.Read API
The API was inconsistent with Writer.Write() and was error-prone.

The new API is consistent with Writer.Write(). Both Writer.Write()
and Reader.Read() now work with a Record that is an exporter field
in the Writer/Reader struct. This makes it obvious that the Record
is owned by the struct. Previously it was not clear if the
pointer to the record returned by Reader.Read() was owned by the
caller and could be modified and what it's lifetime was.
Now it is more intuitively clear that the caller does not own the
record and should not touch it and that the next Read() operation
is going to overwrite it.

Also added pkg.ReadOptions to support reading options. The struct
does not have any options yet, but options are coming in a future PR.