Begin implementing STEF IDL (#43)
This adds formal support for the IDL that is used in the specification.
I plan to eventually replace the JSON representation by this IDL. It is
much easier to use for humans. For example the Otel STEF IDL is 151 easily
readable lines, whereas the equivalent in JSON is 487 harder to read lines.

The IDL will only be used by the generator. The lexer/parser is not subject
to inputs that may come malicious sources, so I am setting the testing and fuzzing
bar relatively low since there is no attack risk and the worse case the generator
will fail to run or will run incorrectly.

The IDL implementation is not yet used anywhere. It is only verified in tests.
Future PRs will replace JSON usage by IDL in the generator.

The parser is a hand-coded recursive descent parser with separate lexer
for tokenization. I looked at alternates, particular at defining the grammar
in ABNF and using ABNF parsers available in Go. Unfortunately none that I tried
worked and it was simpler to just hand code it.

A couple additional changes that I had to do at the same time for simplicity:
- Rename "MainStruct" to "Root" in generator (for naming consistency).
- Change examples in specification.md to use lower-case `key` and `value` (since these are keywords in IDL).