Switch generator to use STEF IDL
The generator now uses STEF IDL format as the input instead of JSON.
I regenerated all serializers and the diff is 0, we got exact match.

There are a few remaining schema tests that still use the JSON
format as testdata. I will clean this up in a future PR, it will
require a bit of refactoring.