Generating enums from IDL (#56)
Resolves https://github.com/splunk/stef/issues/51

We now have separate "storage" and "exported" types
defined for STEF data types. For most types they are
equivalent except for Enums where the "storage" type
is uint64, while "exported" type is user defined.

The generated code now uses either "storage" or "exported"
type, depending on the context and converts between the
types as needed.

The Metric.Type field now uses enum code generation. Previously
manually created enums.go is deleted. enums.go also contains
MetricFlags which is not an enum that we want to generate so
it was moved to internal/metricflags.go.

I will need to refactor internal/metricflags.go, but will do it
in a future PR to keep the size of this PR manageable.

The generator needs more tests to verify how it works with enums
in other data types (oneof,array,multimap). We need to add a better
test suite for generator in the future.