Add basic STEF receiver implementation (#50)
Added ConvertTillEndOfFrame() converter that ensures that we
won't read a record and keep it in memory while blocked on
further reads. Instead all read data is pushed to the pipeline
by STEF receiver quickly after decoding. To make this possible,
introduced a new TillEndOfFrame ReadOption.

I upgraded to new Collector version, that's why lots of go.mod/god.sum
changes and fixes in the stefexporter implementation due to breaking
API changes in the Collector.

Used Copilot to write a small portion of ConvertTillEndOfFrame().

Tested implementation with STEF exporter from Collector:
https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/exporter/stefexporter