Begin implementing Java version
Converted Go's pkg package to Java primarily using Copilot, manually
fixing the most obvious problems. The code is not tested at all.
I am leaving proper testing to the next steps when I will start using
the code in the Java's version of generated serializers.

The only parts that are worth reviewing at the moment is the general
structure of packages, naming conventions, the gradle project
settings and other project-wide machinery.

The implementation details in the methods likely contain bugs and are
not worth manual review at the moment.

The purpose of this PR is to have a first draft of callable methods
and APIs available to help write the code for the next steps.