Breaking: indicate root struct when establishing gRPC connection (#62)
Resolves https://github.com/splunk/stef/issues/58

This will allow the same gRPC destination to accept STEF data
that uses multiple different schemas (one schema per connection).
A known use cases is Otel where we want to send metrics, traces,
logs, etc. This should go to the same gRPC endpoint for
simplicity but the destination needs to know what schema to use
for decoding.

The root struct is indicated in the first message from gRPC
sender to destination, before the STEF stream bytes are sent.