Correctly handle reading STEF with old schema missing a struct
Fixes https://github.com/splunk/stef/issues/86

When the new version of the schema adds a field to oneof and the
type of the field is a new struct that is not present in the old
version of the schema the initialization of the Reader fails since
the decoder tries to find the new struct by the name in the old
schema and fails. Instead of failing the Reader must skip the
new struct and skip initializing its decoder.