Add Go<->Java interoperability tests (#118)
The tests verify STEF implementation interoperability between Go
and Java implementations on a variety of test schemas using
random-generated data.

The tests caught 2 actual bug:
1. In Java implementation the copyFrom function of Multimap with
primitive key and value types was incorrectly copying object references
instead of copying values of key/value fields. See the bug fix in
multimap.java.tmpl.

2. In both Go and Java implementation of structs with optional fields
the copy method was not correctly marking the fields modified.
See the bug fixes in struct.go.tmpl and struct.java.tmpl.

The interop tests are initiated in Java, see toolsTest.java.tmpl file
and the Go portion of the tests is implemented in toolsTest.go.tmpl file,
see comments in toolsTest.java.tmpl that explain how the tests work.

The actual Go and Java files for test schemas is generated during the
build, but is not checked in to git to avoid very large commits.
To verify manually you can run `make` in stefgen/generator (and you can
see the generated interop tests after that), then run `./gradlew build`
in java directory.

Also deleted cmd/diff which we initially planned to implement
these interop tests, but which is no longer needed so we are
using a different approach now.

Resolves https://github.com/splunk/stef/issues/78