Fix oneof encoding bugs
Fixes https://github.com/splunk/stef/issues/130
Fixes https://github.com/splunk/stef/issues/129

The problem was that oneof reset was not fully resetting
the state of the oneof and when the type of the oneof was
switched back it would appear in incorrect state.

This is now fixed by fixing the reset() and setType()
methods.