Correctly encode optional field mask
Resolves https://github.com/splunk/stef/issues/207

The count of optional fields is calculated based on the
overridden schema, like it is done for the count of all
fields. Previously the count of optional fields was static
and that was not working correctly when schema was overridden
schema.

We don't yet have automated tests that verify this fully.
For that we need to generate code for overridden schema.
It will be added in a future PR.