Implement randomized schema test (#227)
Resolves https://github.com/splunk/stef/issues/183

This adds a new TestRandomizedSchema that generates a schema using
random number generator, then compiles and runs the tests for the schema.

The new test revealed a number of bugs. I fixed a few of the smaller ones.
There are more bugs, but I don't want to overload this PR. For now
TestRandomizedSchema does not fail the build when it finds a buggy schema,
but merely logs it. After this PR is merged I will fix every failing
schema one by one.