Remember seeds of failing test cases and rerun
The "seeds" directory contains files with list of seeds for each
test cases that were previously found to generate failures.

The seeds will be re-tested on corresponding test run to ensure
there are no regressions.

If a new random seed is found to cause a failure it will be
automatically appended to the appropriate file.

We will add seeds to the files as we find future bugs and fix them.