Add decode allocation size limit
Contributes to https://github.com/splunk/stef/issues/305

All allocation sizes are now tracked. During decoding the cumulative
allocation size for the current record is checked against a limit
and decoding is aborted if the limit is exceeded.

This fixes numerous bugs in places where for example the decoder
decode a very large length for an array length and tries to allocate
that array.

The bug was caught by fuzzer and corresponding regressions fuzz test
is included in the commit.